Updated content text and changed commit steps
I want to add a few more posts before I continue submitting resumes. It has been a while and I forgot how I prepared the site before. I found a few commands, which had different effects on the codebase. These seem to be the correct commands to run before deployment, so I updated the README with the correct information on what I did.